* bcc32/{Makefile.sub,configure.bat,setup.mak: configure_args   support.
* bcc32/setup.mak: check runtime version.

* win32/win32.c (rb_w32_open_osfhandle): prototype has changed
  in bcc 5.82.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@12831 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ win32/win32.c
@@ -1771,7 +1771,7 @@ init_stdhandle(void)
 
 #ifdef __BORLANDC__
 static int
-rb_w32_open_osfhandle(long osfhandle, int flags)
+rb_w32_open_osfhandle(intptr_t osfhandle, int flags)
 {
     int fd = _open_osfhandle(osfhandle, flags);
     if (fd == -1) {
